APIStorage/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
infi.clickhouse-orm==2.1.3
iso8601==1.1.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
openapi-codec==1.3.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyTelegramBotAPI==4.8.0
pytz==2022.6
requests==2.28.1
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
statsd==4.0.1
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
python-dotenv